Unsure why feed_watchdog was saved into a float instead of a uint32_t since HALGetTick returns uint32_t

+++ src/esw/fw/science/Core/Inc/heater.hpp
@@ -40,7 +40,7 @@ namespace mrover {
         Pin m_heater_pin;
         bool m_state {};
         bool m_auto_shutoff_enabled {};
-        float m_last_time_received_message {};
+        uint32_t m_last_time_received_message {};
         float m_max_heater_temp {DEFAULT_MAX_HEATER_TEMP};
     };
